Add 104-fibonacci.c

+++ functions_nested_loops/104-fibonacci.c
@@ -6,14 +6,14 @@
  */
 int main(void)
 {
-	 unsigned long long int n_1 = 0, n_2 = 1, new, c;
+	 unsigned long int n_1 = 0, n_2 = 1, new, c;
 
 	for (c = 0; c <= 97; c++)
 	{
 		new = n_1 + n_2;
 		n_1 = n_2;
 		n_2 = new;
-		printf("%llu", new);
+		printf("%lu", new);
 		if (c != 97)
 		{
 			printf(", ");
